+++ tools/BugReportTool/BugReportTool/BugReportTool.vcxproj
@@ -38,6 +38,7 @@
     </ClCompile>
     <ClCompile Include="ReportMonitorInfo.cpp" />
     <ClCompile Include="Main.cpp" />
+    <ClCompile Include="ReportRegistry.cpp" />
     <ClCompile Include="ZipTools\zipfolder.cpp" />
   </ItemGroup>
   <ItemGroup>
@@ -53,6 +54,7 @@
     <ClInclude Include="..\..\..\deps\cziplib\src\zip.h" />
     <ClInclude Include="ReportMonitorInfo.h" />
     <ClInclude Include="..\..\..\common\utils\json.h" />
+    <ClInclude Include="ReportRegistry.h" />
     <ClInclude Include="ZipTools\zipfolder.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
