+++ tools/BugReportTool/BugReportTool/ReportMonitorInfo.cpp
@@ -1,56 +1,82 @@
 #pragma once
 #include "ReportMonitorInfo.h"
 #include <Windows.h>
+#include <filesystem>
 #include "../../../src/common/utils/winapi_error.h"
+using namespace std;
 
-int report(std::wostream& os)
+namespace
 {
-    struct capture
+    int buildMonitorInfoReport(std::wostream& os)
     {
-        std::wostream* os = nullptr;
-    };
-
-    auto callback = [](HMONITOR monitor, HDC, RECT*, LPARAM prm) -> BOOL {
-        std::wostream& os = *((capture*)prm)->os;
-        MONITORINFOEX mi;
-        mi.cbSize = sizeof(mi);
-
-        if (GetMonitorInfoW(monitor, &mi))
+        struct capture
         {
-            os << "GetMonitorInfo OK\n";
-            DISPLAY_DEVICE displayDevice = { sizeof(displayDevice) };
+            std::wostream* os = nullptr;
+        };
 
-            DWORD i = 0;
-            while (EnumDisplayDevicesW(mi.szDevice, i++, &displayDevice, EDD_GET_DEVICE_INTERFACE_NAME))
+        auto callback = [](HMONITOR monitor, HDC, RECT*, LPARAM prm) -> BOOL {
+            std::wostream& os = *((capture*)prm)->os;
+            MONITORINFOEX mi;
+            mi.cbSize = sizeof(mi);
+
+            if (GetMonitorInfoW(monitor, &mi))
             {
-                const bool active = displayDevice.StateFlags & DISPLAY_DEVICE_ACTIVE;
-                const bool mirroring = displayDevice.StateFlags & DISPLAY_DEVICE_MIRRORING_DRIVER;
-                os << "EnumDisplayDevices OK:\n"
-                   << "\tMirroring = " << mirroring << '\n'
-                   << "\tActive = " << active << '\n'
-                   << "\tDeviceID = " << displayDevice.DeviceID << '\n'
-                   << "\tDeviceKey = " << displayDevice.DeviceKey << '\n'
-                   << "\tDeviceName = " << displayDevice.DeviceName << '\n'
-                   << "\tDeviceString = " << displayDevice.DeviceString << '\n';
+                os << "GetMonitorInfo OK\n";
+                DISPLAY_DEVICE displayDevice = { sizeof(displayDevice) };
+
+                DWORD i = 0;
+                while (EnumDisplayDevicesW(mi.szDevice, i++, &displayDevice, EDD_GET_DEVICE_INTERFACE_NAME))
+                {
+                    const bool active = displayDevice.StateFlags & DISPLAY_DEVICE_ACTIVE;
+                    const bool mirroring = displayDevice.StateFlags & DISPLAY_DEVICE_MIRRORING_DRIVER;
+                    os << "EnumDisplayDevices OK:\n"
+                        << "\tMirroring = " << mirroring << '\n'
+                        << "\tActive = " << active << '\n'
+                        << "\tDeviceID = " << displayDevice.DeviceID << '\n'
+                        << "\tDeviceKey = " << displayDevice.DeviceKey << '\n'
+                        << "\tDeviceName = " << displayDevice.DeviceName << '\n'
+                        << "\tDeviceString = " << displayDevice.DeviceString << '\n';
+                }
             }
+            else
+            {
+                auto message = get_last_error_message(GetLastError());
+                os << "GetMonitorInfo FAILED: " << (message.has_value() ? message.value() : L"") << '\n';
+            }
+            return TRUE;
+        };
+        capture c;
+        c.os = &os;
+        if (EnumDisplayMonitors(nullptr, nullptr, callback, (LPARAM)&c))
+        {
+            os << "EnumDisplayMonitors OK\n";
         }
         else
         {
             auto message = get_last_error_message(GetLastError());
-            os << "GetMonitorInfo FAILED: " << (message.has_value() ? message.value() : L"") << '\n';
+            os << "EnumDisplayMonitors FAILED: " << (message.has_value() ? message.value() : L"") << '\n';
         }
-        return TRUE;
-    };
-    capture c;
-    c.os = &os;
-    if (EnumDisplayMonitors(nullptr, nullptr, callback, (LPARAM)&c))
-    {
-        os << "EnumDisplayMonitors OK\n";
+        return 0;
     }
-    else
+}
+
+void reportMonitorInfo(const filesystem::path& tmpDir)
+{
+    auto monitorReportPath = tmpDir;
+    monitorReportPath.append("monitor-report-info.txt");
+
+    try
     {
-        auto message = get_last_error_message(GetLastError());
-        os << "EnumDisplayMonitors FAILED: " << (message.has_value() ? message.value() : L"") << '\n';
+        wofstream monitorReport(monitorReportPath);
+        monitorReport << "GetSystemMetrics = " << GetSystemMetrics(SM_CMONITORS) << '\n';
+        buildMonitorInfoReport(monitorReport);
     }
-    return 0;
+    catch (std::exception& ex)
+    {
+        printf("Failed to report monitor info. %s\n", ex.what());
+    }
+    catch (...)
+    {
+        printf("Failed to report monitor info\n");
+    }
 }
