+++ tools/BugReportTool/BugReportTool/ReportRegistry.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <vector>
+#include <string>
+#include <fstream>
+#include <filesystem>
+#include <unordered_map>
+#include <Windows.h>
+
+void reportRegistry(const std::filesystem::path& tmpDir);
